pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "valaw"
version = "0.1.11.dev1"
description = "An asynchronous API wrapper for VALORANT's API"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Jet612" }]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
]
dependencies = [
    "dataclass_wizard==0.39.1",
    "aiohttp==3.13.3",
]

[project.urls]
Source = "https://github.com/Jet612/valaw"
Documentation = "https://valaw.readthedocs.io"
"Issue Tracker" = "https://github.com/Jet612/valaw/issues"
"Chat/Support" = "https://discord.gg/mVXpvunBbF"

[project.optional-dependencies]
dev = ["python-dotenv"]

[tool.setuptools.packages.find]
where = ["src"]
